Seperate the code out from previous merge.
